Central Theme: Cast timestamp to int to avoid PHP notices.
date() expects the timestamp to be an integer, but we're storing it as a string.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org@1137 74240141-8908-4e6f-9713-ba540dce6ec7

+++ public_html/wp-content/themes/wordcamp-central-2012/functions.php
@@ -618,7 +618,7 @@ class WordCamp_Central_Theme {
 		$grouped_wordcamps = array();
 
 		foreach ( $wordcamps as $wordcamp ) {
-			if ( $year = date( 'Y', get_post_meta( $wordcamp->ID, 'Start Date (YYYY-mm-dd)', true ) ) ) {
+			if ( $year = date( 'Y', (int) get_post_meta( $wordcamp->ID, 'Start Date (YYYY-mm-dd)', true ) ) ) {
 				$grouped_wordcamps[ $year ][] = $wordcamp;
 			}
 		}
